Handle titles ordinals and aliases
Letterboxd calls it "Lupin the Third: The Castle of Cagliostro" and
Rotten Tomatoes calls just it "The Castle of Cagliostro". A pain to
handle, but there is an "aka" array in the RT results which includes the
title "Lupin the 3rd: The Castle of Cagliostro". If we convert
ordinals to words and look at titles in the aka array, we can work
around this mismatch.