Bump year range up to 10
This is nuts. Letterboxd says *Now and Then, Here and There* came out in
1999 but Rotten Tomatoes says 2008! I did finally look up this one, and
Letterboxd is right. No idea how RT manages to be off by nearly a
decade. Though Letterboxd says it is a movie and it is actually a
mini-series, so I guess the lesson here is messy data is messy.