Fix everything becoming an ordinal
Turns out the toWords function has an undocumented second parameter,
a boolean flag which creates ordinal words. Since replacers get passed
the index of the match as the second parameter, any match that wasn't at
the start of the string would get converted into an ordinal word.

This doesn't actually affect the matches for the most part, but we'll
tidy it up just in case.

The toOrdinalWords does *not* have an undocumented flag, but we'll
convert that one to an arrow function as well for consistencies' sake.